\name{gfimm-package}
\alias{gfimm-package}
\docType{package}
\title{
\packageTitle{gfimm}
}
\description{
\packageDescription{gfimm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gfimm}
\packageIndices{gfimm}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gfimm}

Maintainer: \packageMaintainer{gfimm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
