% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gfimm}
\alias{gfimm}
\title{Generalized fiducial inference}
\usage{
gfimm(dat, FE, RE, N, thresh = N/2)
}
\arguments{
\item{dat}{data}

\item{FE}{design matrix of fixed effects}

\item{RE}{random effects}

\item{N}{desired number of simulations}

\item{thresh}{threshold}
}
\value{
A list with two elements:
}
\description{
Generalized fiducial inference
}
\examples{
h <- 0.01
dat <- data.frame(npk$yield-h, npk$yield+h)
FE <- as.matrix(rep(1,nrow(dat))) # intercept
RE <- data.frame(block = npk$block)
gfi <- gfimm(dat, FE, RE, N=500)
}
