% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gfiSummary}
\alias{gfiSummary}
\title{Summary of fiducial distributions}
\usage{
gfiSummary(gfi, conf = 0.95)
}
\arguments{
\item{gfi}{Output of \code{\link{gfimm}}}

\item{conf}{confidence level}
}
\value{
Summary statistics in a matrix.
}
\description{
Summary of fiducial distributions
}
\examples{
data(KM41)
h <- 0.005
dat <- cbind(KM41$y-h, KM41$y+h)
FE <- as.matrix(rep(1,nrow(dat))) # intercept
RE <- data.frame(Batch = KM41$Batch)
gfi <- gfimm(dat, FE, RE, N=500)
gfiSummary(gfi)
}
